Fixed all lint issues
