+++ lib/zones/canjs.js
@@ -12,7 +12,7 @@ module.exports = function(){
 
 				cleanup(window.document, setDocument, domMutate);
 			}
-		}
+		};
 	};
 };
 
@@ -32,4 +32,4 @@ function cleanup(document, setDocument, domMutate){
 
 	domMutate.removeChild.call(docEl, body);
 	setDoc(curDoc);
-};
+}
